refactor(frontend): split ActivityPage below 600-line cap
Extract RecentIngestionsTable, MemberActivitySection, and
RecentQueriesSection into components/activity/. Move shared
utilities (formatTimestamp, buildDailyEventCounts) to utils.ts.
ActivityPage.tsx drops from 645 → 272 lines.
