+++ frontend/src/pages/ActivityPage.tsx
@@ -14,21 +14,14 @@ import {
   useAuditEvents,
   useRecentIngestions,
   useInvalidateRecentIngestions,
-  type AuditEventItem,
-  type RecentIngestionRun,
 } from "@/api";
 import { PageContainer } from "@/components/repos/PageContainer";
 import { useEventStream } from "@/hooks/useEventStream";
 import { formatApiError } from "@/lib/errors/api";
-import {
-  ChartSkeleton,
-  formatTimestamp,
-  buildDailyEventCounts,
-} from "./ActivityPage.utils";
-
-// ---------------------------------------------------------------------------
-// ActivityPage — entry point
-// ---------------------------------------------------------------------------
+import { RecentIngestionsTable } from "@/components/activity/RecentIngestionsTable";
+import { MemberActivitySection } from "@/components/activity/MemberActivitySection";
+import { RecentQueriesSection } from "@/components/activity/RecentQueriesSection";
+import { buildDailyEventCounts } from "@/components/activity/utils";
 
 export function ActivityPage(): JSX.Element {
   const me = useMe({ retry: false });
@@ -55,10 +48,6 @@ export function ActivityPage(): JSX.Element {
   return <ActivityPageInner tenantId={me.data.current_tenant.id} />;
 }
 
-// ---------------------------------------------------------------------------
-// Inner page (tenant resolved)
-// ---------------------------------------------------------------------------
-
 interface ActivityPageInnerProps {
   readonly tenantId: string;
 }
@@ -75,7 +64,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
 
   const { events, readyState } = useEventStream(`${apiBase}/v1/ingest/events`);
 
-  // Refetch recent ingestions when a succeeded ingest SSE event arrives (AC5)
   useEffect(() => {
     const latestIngest = events
       .filter((e) => e.type === "ingest.status")
@@ -91,13 +79,11 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
     }
   }, [events, tenantId, invalidateIngestions]);
 
-  // Build audit-events-per-day data for the chart
   const auditChartData = useMemo(() => {
     if (!allAudit.data) return [];
     return buildDailyEventCounts(allAudit.data.events, 14);
   }, [allAudit.data]);
 
-  // Filter member (user) activity client-side
   const memberEvents = useMemo(() => {
     if (!userAudit.data) return [];
     return userAudit.data.events.filter((e) => e.actor_kind === "user");
@@ -105,7 +91,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
 
   return (
     <div className="container max-w-5xl py-8 space-y-8">
-      {/* Page header + SSE status */}
       <header className="flex flex-col gap-1">
         <div className="flex items-center justify-between">
           <h1 className="text-2xl font-semibold tracking-tight">Activity</h1>
@@ -116,7 +101,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
         </p>
       </header>
 
-      {/* Summary cards */}
       <SummaryCards
         repoCount={repos.data?.repos.length ?? 0}
         auditTotal={allAudit.data?.total ?? 0}
@@ -124,7 +108,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
         isLoading={repos.isLoading || allAudit.isLoading}
       />
 
-      {/* Activity over time chart */}
       <section aria-labelledby="chart-heading">
         <h2
           id="chart-heading"
@@ -166,7 +149,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
         </div>
       </section>
 
-      {/* Recent ingestions table */}
       <section aria-labelledby="ingestions-heading">
         <h2
           id="ingestions-heading"
@@ -182,7 +164,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
         />
       </section>
 
-      {/* Member activity */}
       <section aria-labelledby="members-heading">
         <h2
           id="members-heading"
@@ -198,7 +179,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
         />
       </section>
 
-      {/* Recent queries */}
       <section aria-labelledby="queries-heading">
         <h2
           id="queries-heading"
@@ -216,10 +196,6 @@ function ActivityPageInner({ tenantId }: ActivityPageInnerProps): JSX.Element {
     </div>
   );
 }
-
-// ---------------------------------------------------------------------------
-// SSE status badge
-// ---------------------------------------------------------------------------
 
 function SseStatusBadge({
   readyState,
@@ -248,10 +224,6 @@ function SseStatusBadge({
     </div>
   );
 }
-
-// ---------------------------------------------------------------------------
-// Summary cards
-// ---------------------------------------------------------------------------
 
 interface SummaryCardsProps {
   readonly repoCount: number;
@@ -289,301 +261,12 @@ function SummaryCards({
   );
 }
 
-// ---------------------------------------------------------------------------
-// Recent ingestions table
-// ---------------------------------------------------------------------------
-
-interface RecentIngestionsTableProps {
-  readonly runs: readonly RecentIngestionRun[];
-  readonly isLoading: boolean;
-  readonly isError: boolean;
-  readonly error: { status: number; body: unknown } | null;
-}
-
-const STATUS_CELL_CLASS: Record<string, string> = {
-  succeeded: "text-green-600 dark:text-green-400",
-  failed: "text-destructive",
-  running: "text-blue-600 dark:text-blue-400",
-  queued: "text-muted-foreground",
-};
-
-function RunStatusCell({ status }: { status: string }): JSX.Element {
-  const cls = STATUS_CELL_CLASS[status] ?? "text-muted-foreground";
-  return <span className={`text-xs font-medium capitalize ${cls}`}>{status}</span>;
-}
-
-function RecentIngestionsTable({
-  runs,
-  isLoading,
-  isError,
-  error,
-}: RecentIngestionsTableProps): JSX.Element {
-  if (isLoading) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">Loading…</p>
-      </div>
-    );
-  }
-
-  if (isError) {
-    const is404 = error?.status === 404;
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">
-          {is404
-            ? "The recent ingestions endpoint is not yet available. Waiting on backend."
-            : formatApiError(error, "Could not load recent ingestions.")}
-        </p>
-        {is404 && (
-          <p className="mt-1 text-xs text-muted-foreground">
-            Tracked in{" "}
-            <a
-              href="https://github.com/jarnura/rustacean/issues/213"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="underline hover:text-foreground"
-            >
-              GitHub #213
-            </a>
-            .
-          </p>
-        )}
-      </div>
-    );
-  }
-
-  if (runs.length === 0) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">No ingestion runs found.</p>
-      </div>
-    );
-  }
-
+function ChartSkeleton(): JSX.Element {
   return (
-    <div className="overflow-x-auto rounded-lg border border-border">
-      <table className="w-full text-sm" aria-label="Recent ingestion runs">
-        <thead className="border-b border-border bg-muted/40">
-          <tr>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Run ID
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Status
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Started
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Finished
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Trace
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {runs.map((run) => (
-            <tr key={run.id} className="border-b border-border last:border-0 hover:bg-muted/20">
-              <td className="px-4 py-2 font-mono text-xs text-muted-foreground">
-                {run.id.slice(0, 8)}…
-              </td>
-              <td className="px-4 py-2">
-                <RunStatusCell status={run.status} />
-              </td>
-              <td className="px-4 py-2 text-xs text-muted-foreground">
-                {run.started_at ? formatTimestamp(run.started_at) : "—"}
-              </td>
-              <td className="px-4 py-2 text-xs text-muted-foreground">
-                {run.finished_at ? formatTimestamp(run.finished_at) : "—"}
-              </td>
-              <td className="px-4 py-2 font-mono text-xs text-muted-foreground">
-                {run.trace_id ? `${run.trace_id.slice(0, 8)}…` : "—"}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
+    <div
+      role="status"
+      aria-label="Loading chart"
+      className="h-[200px] animate-pulse rounded bg-muted"
+    />
   );
 }
-
-// ---------------------------------------------------------------------------
-// Member activity section
-// ---------------------------------------------------------------------------
-
-interface MemberActivitySectionProps {
-  readonly events: ReadonlyArray<AuditEventItem>;
-  readonly isLoading: boolean;
-  readonly isError: boolean;
-  readonly error: { status: number; body: unknown } | null;
-}
-
-function MemberActivitySection({
-  events,
-  isLoading,
-  isError,
-  error,
-}: MemberActivitySectionProps): JSX.Element {
-  if (isLoading) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">Loading…</p>
-      </div>
-    );
-  }
-
-  if (isError) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">
-          {formatApiError(error, "Could not load member activity.")}
-        </p>
-      </div>
-    );
-  }
-
-  if (events.length === 0) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">No user activity yet.</p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="overflow-x-auto rounded-lg border border-border">
-      <table className="w-full text-sm" aria-label="Member activity events">
-        <thead className="border-b border-border bg-muted/40">
-          <tr>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Action
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Outcome
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Occurred at
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {events.map((event) => (
-            <tr
-              key={event.id}
-              className="border-b border-border last:border-0 hover:bg-muted/20"
-            >
-              <td className="px-4 py-2 font-mono text-xs">{event.action}</td>
-              <td className="px-4 py-2">
-                <span
-                  className={`text-xs font-medium ${
-                    event.outcome === "success"
-                      ? "text-green-600 dark:text-green-400"
-                      : "text-destructive"
-                  }`}
-                >
-                  {event.outcome}
-                </span>
-              </td>
-              <td className="px-4 py-2 text-xs text-muted-foreground">
-                {formatTimestamp(event.occurred_at)}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Recent queries section
-// ---------------------------------------------------------------------------
-
-interface RecentQueriesSectionProps {
-  readonly events: ReadonlyArray<AuditEventItem>;
-  readonly isLoading: boolean;
-  readonly isError: boolean;
-  readonly error: { status: number; body: unknown } | null;
-}
-
-function RecentQueriesSection({
-  events,
-  isLoading,
-  isError,
-  error,
-}: RecentQueriesSectionProps): JSX.Element {
-  if (isLoading) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">Loading…</p>
-      </div>
-    );
-  }
-
-  if (isError) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">
-          {formatApiError(error, "Could not load recent queries.")}
-        </p>
-      </div>
-    );
-  }
-
-  if (events.length === 0) {
-    return (
-      <div className="rounded-lg border border-border bg-card p-4">
-        <p className="text-sm text-muted-foreground">
-          No search queries recorded yet.
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="overflow-x-auto rounded-lg border border-border">
-      <table className="w-full text-sm" aria-label="Recent search queries">
-        <thead className="border-b border-border bg-muted/40">
-          <tr>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Action
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Outcome
-            </th>
-            <th scope="col" className="px-4 py-2 text-left font-medium text-muted-foreground">
-              Occurred at
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {events.map((event) => (
-            <tr
-              key={event.id}
-              className="border-b border-border last:border-0 hover:bg-muted/20"
-            >
-              <td className="px-4 py-2 font-mono text-xs">{event.action}</td>
-              <td className="px-4 py-2">
-                <span
-                  className={`text-xs font-medium ${
-                    event.outcome === "success"
-                      ? "text-green-600 dark:text-green-400"
-                      : "text-destructive"
-                  }`}
-                >
-                  {event.outcome}
-                </span>
-              </td>
-              <td className="px-4 py-2 text-xs text-muted-foreground">
-                {formatTimestamp(event.occurred_at)}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
